requirements.txt
pytest
mysql.connector
mysql-connector-python